backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "fastapi-login>=1.10.3",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "momento>=1.27.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-decouple>=3.8",
    "python-multipart>=0.0.20",
]
